+++ package.json
@@ -9,15 +9,21 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@react-navigation/bottom-tabs": "^7.2.0",
+    "@react-navigation/native": "^7.0.14",
     "expo": "~52.0.20",
+    "expo-constants": "~17.0.3",
+    "expo-linking": "~7.0.3",
+    "expo-router": "^4.0.15",
     "expo-status-bar": "~2.0.0",
+    "nodemon": "^3.1.9",
     "react": "18.3.1",
     "react-native": "0.76.5",
-    "expo-router": "~4.0.14",
-    "react-native-safe-area-context": "4.12.0",
-    "react-native-screens": "~4.1.0",
-    "expo-linking": "~7.0.3",
-    "expo-constants": "~17.0.3"
+    "react-native-gesture-handler": "^2.20.2",
+    "react-native-reanimated": "^3.16.6",
+    "react-native-safe-area-context": "^4.12.0",
+    "react-native-screens": "^4.4.0",
+    "react-native-vector-icons": "^10.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
